myblog/apps/users/models.py
from datetime import datetime

from django.db import models
from django.contrib.auth.models import AbstractUser
from imagekit.models import ProcessedImageField
from imagekit.processors import ResizeToFill


class UserProfile(AbstractUser):
    """
    用户信息
    """
    link = models.URLField(
        '个人网站', blank=True, help_text='提示：网址必须是以http开头的完整形式')
    avatar = ProcessedImageField(upload_to='avatar/%Y/%m/%d', default='avatar/default.png',
                                 verbose_name='头像', processors=[ResizeToFill(80, 80)])
    gender = models.CharField(max_length=6, choices=(("male", u"男"), ("female", "女")), default="female", verbose_name="性别")
    mobile = models.CharField(null=True, blank=True, max_length=11, verbose_name="电话")

    class Meta:
        verbose_name = '用户'
        verbose_name_plural = verbose_name
        ordering = ['-id']
    
    def __str__(self):
        return self.username

class VerifyCode(models.Model):
    """
    短信验证码
    """
    code = models.CharField(max_length=10, verbose_name="验证码")
    mobile = models.CharField(max_length=11, verbose_name="电话")
    add_time = models.DateTimeField(default=datetime.now, verbose_name="添加时间")

    class Meta:
        verbose_name = "短信验证码"
        verbose_name_plural = verbose_name

    def __str__(self):
        return self.code